refactor(oas): subconverters oas2/oas3 subclasses
relates-to #120
