+++ packages/oas/src/converter/SubPart.ts
