+++ packages/oas/src/converter/parts/BaseUrlParser.ts
@@ -1,5 +1,5 @@
-import { isOASV2, isOASV3 } from '../utils';
-import { ConvertError } from '../errors';
+import { isOASV2, isOASV3 } from '../../utils';
+import { ConvertError } from '../../errors';
 import { Sampler } from './Sampler';
 import { UriTemplator } from './UriTemplator';
 import {
