+++ packages/oas/src/converter/parts/headers/index.ts
@@ -1,0 +1,2 @@
+export { Oas2HeadersConverter } from './Oas2HeadersConveter';
+export { Oas3HeadersConverter } from './Oas3HeadersConverter';
