+++ packages/oas/src/converter/parts/path/index.ts
@@ -1,0 +1,2 @@
+export { Oas2PathConverter } from './Oas2PathConverter';
+export { Oas3PathConverter } from './Oas3PathConverter';
