--- packages/oas/src/converter/subconverters/index.ts
@@ -1,6 +1,0 @@
-export { HeadersConverter } from './HeadersConverter';
-export { PathConverter } from './PathConverter';
-export { PostDataConverter } from './PostDataConverter';
-export { QueryStringConverter } from './QueryStringConverter';
-export { SubConverter } from './SubConverter';
-export { SubPart } from './SubPart';
